datetime.from_excel(): round datetimes to full milliseconds
According to the specification (section 18.17.4), fractional seconds in
datetimes are stored with up to three decimals, which corresponds to
millisecond precision. It does not make sense to pretend we could be
more accurate than that.

The condition for detecting desired time-only returns needs to be
adapted because we might round up to a full day and should not return
time only in that case.

--HG--
branch : 3.0